+++ commands/node_vote_test.go
@@ -11,8 +11,6 @@ import (
 
 func TestCheckNodeVote(t *testing.T) {
 	t.Run("Submitting a nil command fails", testNilNodeVoteFails)
-	t.Run("Submitting a node vote without pub key fails", testNodeVoteWithoutPubKeyFails)
-	t.Run("Submitting a node vote with pub key succeeds", testNodeVoteWithPubKeySucceeds)
 	t.Run("Submitting a node vote without reference fails", testNodeVoteWithoutReferenceFails)
 	t.Run("Submitting a node vote with reference succeeds", testNodeVoteWithReferenceSucceeds)
 }
@@ -21,18 +19,6 @@ func testNilNodeVoteFails(t *testing.T) {
 	err := checkNodeVote(nil)
 
 	assert.Error(t, err)
-}
-
-func testNodeVoteWithoutPubKeyFails(t *testing.T) {
-	err := checkNodeVote(&commandspb.NodeVote{})
-	assert.Contains(t, err.Get("node_vote.pub_key"), commands.ErrIsRequired)
-}
-
-func testNodeVoteWithPubKeySucceeds(t *testing.T) {
-	err := checkNodeVote(&commandspb.NodeVote{
-		PubKey: []byte("0xDEADBEEF"),
-	})
-	assert.NotContains(t, err.Get("node_vote.pub_key"), commands.ErrIsRequired)
 }
 
 func testNodeVoteWithoutReferenceFails(t *testing.T) {